feat: 模板模式串联抽奖规则

+++ big-market-huling-domain/src/main/java/com/huling/domain/strategy/service/rule/tree/ILogicTreeNode.java
@@ -4,6 +4,6 @@ import com.huling.domain.strategy.service.rule.tree.factory.DefaultTreeFactory;
 
 public interface ILogicTreeNode {
 
-    DefaultTreeFactory.TreeActionEntity logic(String userId, Long strategyId, Integer awardId);
+    DefaultTreeFactory.TreeActionEntity logic(String userId, Long strategyId, Long awardId);
 
 }
