+++ big-market-huling-domain/src/main/java/com/huling/domain/strategy/service/rule/tree/factory/engine/IDecisionTreeEngine.java
@@ -4,6 +4,6 @@ import com.huling.domain.strategy.service.rule.tree.factory.DefaultTreeFactory;
 
 public interface IDecisionTreeEngine {
 
-    DefaultTreeFactory.StrategyAwardVO process(String userId, Long strategyId, Integer awardId);
+    DefaultTreeFactory.StrategyAwardVO process(String userId, Long strategyId, Long awardId);
 
 }
