+++ big-market-huling-domain/src/main/java/com/huling/domain/strategy/service/rule/tree/factory/engine/impl/DecisionTreeEngine.java
@@ -23,7 +23,7 @@ public class DecisionTreeEngine implements IDecisionTreeEngine {
     }
 
     @Override
-    public DefaultTreeFactory.StrategyAwardVO process(String userId, Long strategyId, Integer awardId) {
+    public DefaultTreeFactory.StrategyAwardVO process(String userId, Long strategyId, Long awardId) {
 
         // 1.获取决策树基本信息
         String currentNode = ruleTreeVO.getTreeRootRuleNode();
